// src/app/modulos/auth/models/register-request.ts
export class RegisterRequest {
    idTipoDocumento: number=0;
    idRol: number=0;
    nombre: string="";
    apellidoPaterno: string="";
    apellidoMaterno: string="";
    numeroDocumento: string="";
    telefono: string="";
    userName: string="";
    password: string="";
    email: string="";
}